Clean install.sh script, add tmux-plugins/tpm installation

+++ install.sh
@@ -38,7 +38,19 @@ else
     warning "You already have some personal fonts, install the vim fonts manually: https://github.com/Lokaltog/powerline-fonts"
 fi
 
-#
+# Install Vundle and vim plugins
+if [ ! -d ~/.vim/bundle/vundle  ]; then
+  echo "Installing Vundle for vim.."
+  git clone https://github.com/gmarik/Vundle.vim ~/.vim/bundle/vundle
+
+  echo "Installing vim plugins..."
+  vim -c BundleInstall -c q
+else
+  cd vim/bundle/vundle;
+  git pull;
+  cd ../../..;
+fi;
+
 # Install powerline configuration
 #
 if [ ! -d ~/.config/powerline ]; then
@@ -48,79 +60,34 @@ else
     warning "There is a powerline configuration folder already installed"
 fi
 
-
-function install_if_does_not_exist {
-    if [ "`cat /tmp/pip_freeze|grep $1|wc -l`" -eq "0" ]; then
-        sudo pip install $1
+# Create symlinks
+files=`ls | sed -e "/\(install.sh\)/d"`
+for file in $files; do
+    if [ $file != "LICENSE" ] && [ ! -f ~/.$file ] && [ ! -d ~/.$file ]; then
+        echo "Symlinking $file..."
+        ln -s $PWD/$file ~/.$file
     fi
-}
-
-# install_if_does_not_exist flake8
-# install_if_does_not_exist ropevim
-# install_if_does_not_exist pyflakes
-
-#
-# Create all the links
-#
-function warning_if_exists {
-    if [ "$?" == 1 ]; then
-        warning "$1 configuration already exists, remove it and try again!"
-    else
-        ok "$1 configurations installed"
-    fi
-}
+done;
 
 # Install shell config files
 ln -s $PWD/zshrc ~/.zshrc &> /dev/null
-warning_if_exists zsh
-
 ln -s $PWD/bashrc ~/.bashrc &> /dev/null
-warning_if_exists bashrc
-
-# Install additional shell config files
 ln -s $PWD/shell-configs/aliases ~/.aliases &> /dev/null
-warning_if_exists aliases
-
-# Install additional shell config files
 ln -s $PWD/shell-configs/profile ~/.profile &> /dev/null
-warning_if_exists profile
-
 ln -s $PWD/shell-configs/functions ~/.functions &> /dev/null
-warning_if_exists functions
-
 ln -s $PWD/shell-configs/exports ~/.exports &> /dev/null
-warning_if_exists exports
-
-# Install git config files
 ln -s $PWD/gitconfig ~/.gitconfig &> /dev/null
-warning_if_exists gitconfig
-
-cp $PWD/gitconfig-additional ~/.gitconfig-additional &> /dev/null
-warning_if_exists gitconfig-additional
-
 ln -s $PWD/gitignore ~/.gitignore &> /dev/null
-warning_if_exists gitignore
-
-# Install multiplexer-shell config files
 ln -s $PWD/screenrc ~/.screenrc &> /dev/null
-warning_if_exists screenrc
-
 ln -s $PWD/tmux.conf ~/.tmux.conf &> /dev/null
-warning_if_exists tmux
-
-# Install editor config files
 ln -s $PWD/vimrc ~/.vimrc &> /dev/null
-warning_if_exists vimrc
 ln -s $PWD/vim ~/.vim &> /dev/null
-warning_if_exists vim
-rm $HOME/.vim/vim  # just remove it, don't even bother checking
-
-if [ ! -d ~/.config/powerline ]; then
-    git clone https://github.com/gmarik/Vundle.vim.git vim/bundle/vundle
-else
-    cd vim/bundle/vundle;
-    git pull;
-    cd ../../..;
-fi
 
 ok "Configurations linked properly"
+
+git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+
+if [ ! -f ~/.gitconfig-additional ]; then
+  cp $PWD/gitconfig-additional ~/.gitconfig-additional &> /dev/null
+fi
+
